refactor: moved api.go to internal root
